Refactor code for consistency and readability across multiple files

+++ script/Interactions.s.sol
@@ -1,12 +1,13 @@
 //SPDX-License-Identifier: MIT
 
 pragma solidity ^0.8.19;
+
 import {Script, console} from "forge-std/Script.sol";
 import {DevOpsTools} from "foundry-devops/src/DevOpsTools.sol";
 import {FundMe} from "../src/FundMe.sol";
 
 contract FundFundMe is Script {
-    uint constant SEND_VALUE = 0.1 ether;
+    uint256 constant SEND_VALUE = 0.1 ether;
 
     function fundFundMe(address mostRecentDeployed) public {
         vm.startBroadcast();
@@ -16,10 +17,7 @@ contract FundFundMe is Script {
     }
 
     function run() external {
-        address mostRecentlyDeployed = DevOpsTools.get_most_recent_deployment(
-            "FundMe",
-            block.chainid
-        );
+        address mostRecentlyDeployed = DevOpsTools.get_most_recent_deployment("FundMe", block.chainid);
         fundFundMe(mostRecentlyDeployed);
     }
 }
@@ -33,10 +31,7 @@ contract WithdrawFundMe is Script {
     }
 
     function run() external {
-        address mostRecentlyDeployed = DevOpsTools.get_most_recent_deployment(
-            "FundMe",
-            block.chainid
-        );
+        address mostRecentlyDeployed = DevOpsTools.get_most_recent_deployment("FundMe", block.chainid);
         withdrawFundMe(mostRecentlyDeployed);
     }
 }
