+++ test/unit/FundMeTest.t.sol
@@ -12,6 +12,7 @@ contract FundMeTest is Test {
     uint256 constant SEND_VALUE = 1 ether;
 
     uint256 constant STARTING_BALANCE = 10 ether;
+
     modifier funded() {
         vm.prank(alice);
         fundMe.fund{value: SEND_VALUE}();
@@ -23,8 +24,9 @@ contract FundMeTest is Test {
         fundMe = deployFundMe.run();
         vm.deal(alice, STARTING_BALANCE);
     }
+
     function testDemo() public pure {
-        assertEq(uint(1337), uint(1337));
+        assertEq(uint256(1337), uint256(1337));
     }
 
     function testOwnerCheck() public {
@@ -39,11 +41,13 @@ contract FundMeTest is Test {
         uint256 version = fundMe.getVersion();
         assertEq(version, 4);
     }
+
     function testFundUpdatesFundDataStructure() public {
         fundMe.fund{value: 10 ether}();
         uint256 amountFunded = fundMe.getAddressToAmountFunded(address(this));
         assertEq(amountFunded, 10 ether);
     }
+
     function testAddsFunderToArrayOfFunders() public {
         vm.startPrank(alice);
         fundMe.fund{value: SEND_VALUE}();
@@ -52,10 +56,12 @@ contract FundMeTest is Test {
         address funder = fundMe.getFunder(0);
         assertEq(funder, alice);
     }
+
     function testOnlyOwnerCanWithdraw() public funded {
         vm.expectRevert();
         fundMe.withdraw();
     }
+
     function testWithdrawFromASingleFunder() public funded {
         uint256 startingFundMeBalance = address(fundMe).balance;
 
@@ -67,19 +73,13 @@ contract FundMeTest is Test {
         uint256 endingFundMeBalance = address(fundMe).balance;
         uint256 endingOwnerBalance = fundMe.getOwner().balance;
         assertEq(endingFundMeBalance, 0);
-        assertEq(
-            startingFundMeBalance + startingOwnerBalance,
-            endingOwnerBalance
-        );
+        assertEq(startingFundMeBalance + startingOwnerBalance, endingOwnerBalance);
     }
+
     function testWithdrawFromMultipleFunders() public funded {
         uint160 numberOfFunders = 10;
         uint160 startingFunderIndex = 1;
-        for (
-            uint160 i = startingFunderIndex;
-            i < numberOfFunders + startingFunderIndex;
-            i++
-        ) {
+        for (uint160 i = startingFunderIndex; i < numberOfFunders + startingFunderIndex; i++) {
             // we get hoax from stdcheats
             // prank + deal
             hoax(address(i), SEND_VALUE);
@@ -94,14 +94,8 @@ contract FundMeTest is Test {
         vm.stopPrank();
 
         assert(address(fundMe).balance == 0);
-        assert(
-            startingFundMeBalance + startingOwnerBalance ==
-                fundMe.getOwner().balance
-        );
-        assert(
-            (numberOfFunders + 1) * SEND_VALUE ==
-                fundMe.getOwner().balance - startingOwnerBalance
-        );
+        assert(startingFundMeBalance + startingOwnerBalance == fundMe.getOwner().balance);
+        assert((numberOfFunders + 1) * SEND_VALUE == fundMe.getOwner().balance - startingOwnerBalance);
     }
 
     // Gas comparison tests
@@ -109,11 +103,7 @@ contract FundMeTest is Test {
         // Setup multiple funders for a meaningful gas comparison
         uint160 numberOfFunders = 10;
         uint160 startingFunderIndex = 1;
-        for (
-            uint160 i = startingFunderIndex;
-            i < numberOfFunders + startingFunderIndex;
-            i++
-        ) {
+        for (uint160 i = startingFunderIndex; i < numberOfFunders + startingFunderIndex; i++) {
             hoax(address(i), SEND_VALUE);
             fundMe.fund{value: SEND_VALUE}();
         }
@@ -132,11 +122,7 @@ contract FundMeTest is Test {
         // Setup multiple funders for a meaningful gas comparison
         uint160 numberOfFunders = 10;
         uint160 startingFunderIndex = 1;
-        for (
-            uint160 i = startingFunderIndex;
-            i < numberOfFunders + startingFunderIndex;
-            i++
-        ) {
+        for (uint160 i = startingFunderIndex; i < numberOfFunders + startingFunderIndex; i++) {
             hoax(address(i), SEND_VALUE);
             fundMe.fund{value: SEND_VALUE}();
         }
